requirements.txt
fastapi
uvicorn
langchain
openai
transformers
sentence-transformers
pinecone-client
psycopg2-binary
sqlalchemy
redis
python-dotenv
requests
pydantic
loguru
